Move flourish playback to own trigger source
